improve rolling_quantile kernel (no nulls) ~28x (#3437)
* improve quantile kernel (no nulls) ~28x
* improve quantile nulls
