+++ polars/polars-arrow/src/kernels/rolling/mod.rs
@@ -1,6 +1,23 @@
 pub mod no_nulls;
 pub mod nulls;
+pub mod quantile_no_nulls;
+pub mod quantile_nulls;
+mod window;
+
+use crate::data_types::IsFloat;
+use crate::prelude::QuantileInterpolOptions;
+use crate::utils::CustomIterTools;
+use arrow::array::{ArrayRef, PrimitiveArray};
+use arrow::bitmap::utils::{count_zeros, get_bit_unchecked};
 use arrow::bitmap::{Bitmap, MutableBitmap};
+use arrow::types::NativeType;
+use num::ToPrimitive;
+use num::{Bounded, Float, NumCast, One, Zero};
+use std::cmp::Ordering;
+use std::ops::AddAssign;
+use std::ops::{Add, Div, Mul, Sub};
+use std::sync::Arc;
+use window::*;
 
 type Start = usize;
 type End = usize;
@@ -56,5 +73,25 @@ where
         Some(validity)
     } else {
         None
+    }
+}
+pub(super) fn sort_buf<T>(buf: &mut [T])
+where
+    T: IsFloat + NativeType + PartialOrd,
+{
+    if T::is_float() {
+        buf.sort_by(|a, b| {
+            match (a.is_nan(), b.is_nan()) {
+                // safety: we checked nans
+                (false, false) => unsafe { a.partial_cmp(b).unwrap_unchecked() },
+                (true, true) => Ordering::Equal,
+                (true, false) => Ordering::Greater,
+                (false, true) => Ordering::Less,
+            }
+        });
+    } else {
+        // Safety:
+        // all integers are Ord
+        unsafe { buf.sort_by(|a, b| a.partial_cmp(b).unwrap_unchecked()) };
     }
 }
